atopa-encuestas/requirements.txt
APScheduler
Flask
SQLAlchemy
redis
PyJWT
Werkzeug
reportlab
html2image
pyvis
Pillow
matplotlib
networkx
Flask-Babel
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
Flask-Mail
Flask-MySQL
Flask-RESTful
Flask-SQLAlchemy
mysqlclient